requirements/app.txt
fastapi
python-multipart
uvicorn
SQLAlchemy
psycopg2
pydantic
numpy
validators
pyyaml
typer
rich